requirements.txt
openai
langchain
langchain-community
streamlit
python-dotenv
pyPDF2
langchain-anthropic
